library management for schools and college will be helpful
this project was impletemented in my learing time of php. so it simple to understand and easy to use.i have given the mysql files in the folder .

i have provide the upload  option for students and book information in bulk order .
100 of members can be uploaded in one click .

i am not applyed bootstrap at that time u can  improve it.

Thank You for choosing  this project .